Chore: Increase Setting type safety
