+++ packages/app-mobile/root.tsx
@@ -14,7 +14,7 @@ import ResourceService from '@joplin/lib/services/ResourceService';
 import KvStore from '@joplin/lib/services/KvStore';
 import NoteScreen from './components/screens/Note';
 import UpgradeSyncTargetScreen from './components/screens/UpgradeSyncTargetScreen';
-import Setting, { Env } from '@joplin/lib/models/Setting';
+import Setting, { AppType, Env } from '@joplin/lib/models/Setting';
 import PoorManIntervals from '@joplin/lib/PoorManIntervals';
 import reducer, { NotesParent, parseNotesParent, serializeNotesParent } from '@joplin/lib/reducer';
 import ShareExtension from './utils/ShareExtension';
@@ -501,9 +501,9 @@ async function initialize(dispatch: Function) {
 		value: profileConfig,
 	});
 
-	Setting.setConstant('env', __DEV__ ? 'dev' : 'prod');
+	Setting.setConstant('env', __DEV__ ? Env.Dev : Env.Prod);
 	Setting.setConstant('appId', 'net.cozic.joplin-mobile');
-	Setting.setConstant('appType', 'mobile');
+	Setting.setConstant('appType', AppType.Mobile);
 	Setting.setConstant('tempDir', await initializeTempDir());
 	Setting.setConstant('cacheDir', `${getProfilesRootDir()}/cache`);
 	const resourceDir = getResourceDir(currentProfile, isSubProfile);
