+++ packages/lib/reducer.ts
@@ -12,6 +12,7 @@ import { getListRendererIds } from './services/noteList/renderers';
 import { ProcessResultsRow } from './services/search/SearchEngine';
 import { getDisplayParentId } from './services/trash';
 import Logger from '@joplin/utils/Logger';
+import { SettingsRecord } from './models/settings/types';
 const fastDeepEqual = require('fast-deep-equal');
 const { ALL_NOTES_FILTER_ID } = require('./reserved-ids');
 const { createSelectorCreator, defaultMemoize } = require('reselect');
@@ -100,8 +101,7 @@ export interface State {
 	syncReport: any;
 	searchQuery: string;
 	searchResults: ProcessResultsRow[];
-	// eslint-disable-next-line @typescript-eslint/no-explicit-any -- Old code before rule was applied
-	settings: Record<string, any>;
+	settings: Partial<SettingsRecord>;
 	// eslint-disable-next-line @typescript-eslint/no-explicit-any -- Old code before rule was applied
 	sharedData: any;
 	appState: string;
